更新数据库站点表数据

+++ code/pbsproject/src/main/java/pbs/base/pojo/po/PbsRentInfo.java
@@ -13,9 +13,9 @@ public class PbsRentInfo {
 
     private Integer maxNumber;
 
-    private Float lat;
+    private Double lat;
 
-    private Float lng;
+    private Double lng;
 
     private String leftOrRight;
 
@@ -75,19 +75,19 @@ public class PbsRentInfo {
         this.maxNumber = maxNumber;
     }
 
-    public Float getLat() {
+    public Double getLat() {
         return lat;
     }
 
-    public void setLat(Float lat) {
+    public void setLat(Double lat) {
         this.lat = lat;
     }
 
-    public Float getLng() {
+    public Double getLng() {
         return lng;
     }
 
-    public void setLng(Float lng) {
+    public void setLng(Double lng) {
         this.lng = lng;
     }
